feat(component): refactor selector
